Increased tick count to long

+++ src/World.h
@@ -773,7 +773,7 @@ private:
 	ScheduledTaskList m_ScheduledTasks;
 	
 	/// Current Tick number for sceduled tasks
-	int m_TickID;
+	long m_TickID;
 	
 	/// Guards m_Clients
 	cCriticalSection  m_CSClients;
